Moved Toggle and Array
